adjust types in SongFormat

+++ src/Common/SongFormat.h
@@ -19,23 +19,23 @@ bool operator != (SongFormat const& lhs, SongFormat const& rhs);
 struct SongFormat
 {
     // specifies the number of samples per seconds
-    unsigned int SampleRate = 0;
+    uint32_t SampleRate = 0;
 
     // specifies the number of (mono) audio channels
-    unsigned int Channels();
+    uint32_t Channels();
 
     // indicates the type of the items in pcm buffer "data" from PCMHolder
     SampleFormat_t SampleFormat = SampleFormat_t::unknown;
 
     // number of voices available for a song
     // voice == group of consecutive audio channels
-    unsigned int Voices = 0;
+    uint16_t Voices = 0;
     
     vector<string> VoiceName;
     
     vector<bool> VoiceIsMuted;
     
-    vector<unsigned int> VoiceChannels;
+    vector<uint16_t> VoiceChannels;
 
     void SetVoices(unsigned int nVoices);
